Pull out Invite Modal into own file
