+++ frontend/src/__generated/scss/pages/WorkspaceTeam/WorkspaceTeam.module.scss.js
@@ -1,1 +1,1 @@
-var t={subTitle:"_subTitle_eqxhx_1",membersTitleContainer:"_membersTitleContainer_eqxhx_5",membersTitle:"_membersTitle_eqxhx_5",memberCard:"_memberCard_eqxhx_13",memberCardWrapper:"_memberCardWrapper_eqxhx_20",userDetails:"_userDetails_eqxhx_26",name:"_name_eqxhx_33",email:"_email_eqxhx_40",role:"_role_eqxhx_48",header:"_header_eqxhx_54",buttonRow:"_buttonRow_eqxhx_59",emailInput:"_emailInput_eqxhx_66",inviteButton:"_inviteButton_eqxhx_70",boxSubTitle:"_boxSubTitle_eqxhx_74",removeTeamMemberButton:"_removeTeamMemberButton_eqxhx_79",titleContainer:"_titleContainer_eqxhx_85",hr:"_hr_eqxhx_98"};export{t as default};
+var r={subTitle:"_subTitle_1hx0w_1",membersTitleContainer:"_membersTitleContainer_1hx0w_5",membersTitle:"_membersTitle_1hx0w_5",memberCard:"_memberCard_1hx0w_13",memberCardWrapper:"_memberCardWrapper_1hx0w_20",userDetails:"_userDetails_1hx0w_26",name:"_name_1hx0w_33",email:"_email_1hx0w_40",role:"_role_1hx0w_48",header:"_header_1hx0w_54",removeTeamMemberButton:"_removeTeamMemberButton_1hx0w_59",titleContainer:"_titleContainer_1hx0w_65"};export{r as default};
