+++ frontend/src/__generated/scss/pages/WorkspaceTeam/components/InviteMemberModal.module.scss.js
@@ -1,0 +1,1 @@
+var o={boxSubTitle:"_boxSubTitle_1g9q5_1",buttonRow:"_buttonRow_1g9q5_6",emailInput:"_emailInput_1g9q5_13",inviteButton:"_inviteButton_1g9q5_17",hr:"_hr_1g9q5_21"};export{o as default};
