+++ frontend/src/pages/WorkspaceTeam/WorkspaceTeam.module.scss
@@ -56,26 +56,6 @@
 	display: flex;
 }
 
-.buttonRow {
-	align-items: center;
-	column-gap: var(--size-medium);
-	display: grid;
-	grid-template-columns: 1fr 120px;
-}
-
-.emailInput {
-	height: 40px;
-}
-
-.inviteButton {
-	width: 100%;
-}
-
-.boxSubTitle {
-	color: var(--color-gray-500) !important;
-	margin-top: 0 !important;
-}
-
 .removeTeamMemberButton {
 	flex-shrink: 0;
 	height: fit-content;
@@ -94,10 +74,4 @@
 		margin-bottom: 0 !important;
 		margin-top: var(--size-xSmall) !important;
 	}
-}
-
-.hr {
-	border: 0;
-	border-top: 1px solid var(--color-gray-300);
-	margin: var(--size-large) 0;
 }
