+++ frontend/src/pages/WorkspaceTeam/WorkspaceTeam.module.scss.d.ts
@@ -1,10 +1,5 @@
-export const boxSubTitle: string
-export const buttonRow: string
 export const email: string
-export const emailInput: string
 export const header: string
-export const hr: string
-export const inviteButton: string
 export const memberCard: string
 export const memberCardWrapper: string
 export const membersTitle: string
