+++ frontend/src/pages/WorkspaceTeam/WorkspaceTeam.tsx
@@ -1,25 +1,18 @@
 import { useAuthContext } from '@authentication/AuthContext'
-import Alert from '@components/Alert/Alert'
 import { AdminAvatar } from '@components/Avatar/Avatar'
 import Card from '@components/Card/Card'
-import CopyText from '@components/CopyText/CopyText'
-import Input from '@components/Input/Input'
-import { CircularSpinner } from '@components/Loading/Loading'
-import Modal from '@components/Modal/Modal'
 import Select from '@components/Select/Select'
 import Table from '@components/Table/Table'
 import {
 	useChangeAdminRoleMutation,
 	useDeleteAdminFromWorkspaceMutation,
 	useGetWorkspaceAdminsQuery,
-	useSendAdminWorkspaceInviteMutation,
 } from '@graph/hooks'
 import { namedOperations } from '@graph/operations'
 import { AdminRole } from '@graph/schemas'
 import { Box } from '@highlight-run/ui'
 import SvgTrashIcon from '@icons/TrashIcon'
 import AutoJoinForm from '@pages/WorkspaceTeam/components/AutoJoinForm'
-import { getWorkspaceInvitationLink } from '@pages/WorkspaceTeam/utils'
 import {
 	Authorization,
 	useAuthorization,
@@ -29,12 +22,11 @@ import { useParams } from '@util/react-router/useParams'
 import { getDisplayNameFromEmail, titleCaseString } from '@util/string'
 import { message } from 'antd'
 import clsx from 'clsx'
-import moment from 'moment'
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import { useToggle } from 'react-use'
-import { StringParam, useQueryParam } from 'use-query-params'
 
-import commonStyles from '../../Common.module.scss'
+import InviteMemberModal from '@/pages/WorkspaceTeam/components/InviteMemberModal'
+
 import Button from '../../components/Button/Button/Button'
 import layoutStyles from '../../components/layout/LeadAlignLayout.module.scss'
 import PopConfirm from '../../components/PopConfirm/PopConfirm'
@@ -42,15 +34,12 @@ import styles from './WorkspaceTeam.module.scss'
 
 const WorkspaceTeam = () => {
 	const { workspace_id } = useParams<{ workspace_id: string }>()
-	const emailRef = useRef<null | HTMLInputElement>(null)
 	const { data, error, loading } = useGetWorkspaceAdminsQuery({
 		variables: { workspace_id: workspace_id! },
 		skip: !workspace_id,
 	})
 	const { checkPolicyAccess } = useAuthorization()
-	const [email, setEmail] = useState('')
 	const [showModal, toggleShowModal] = useToggle(false)
-	const [newAdminRole, setNewAdminRole] = useState<AdminRole>(AdminRole.Admin)
 
 	const { admin } = useAuthContext()
 	const [deleteAdminFromWorkspace] = useDeleteAdminFromWorkspaceMutation({
@@ -71,41 +60,6 @@ const WorkspaceTeam = () => {
 	})
 	const [changeAdminRole] = useChangeAdminRoleMutation()
 
-	const [autoinvite_email] = useQueryParam('autoinvite_email', StringParam)
-
-	useEffect(() => {
-		if (autoinvite_email) {
-			setEmail(autoinvite_email)
-			setNewAdminRole(AdminRole.Member)
-			toggleShowModal(true)
-		}
-	}, [autoinvite_email, toggleShowModal])
-
-	const [
-		sendInviteEmail,
-		{ loading: sendLoading, data: sendInviteEmailData },
-	] = useSendAdminWorkspaceInviteMutation()
-
-	const onSubmit = (e: { preventDefault: () => void }) => {
-		e.preventDefault()
-
-		if (!workspace_id) {
-			return
-		}
-
-		sendInviteEmail({
-			variables: {
-				workspace_id: workspace_id!,
-				email,
-				base_url: window.location.origin,
-				role: newAdminRole,
-			},
-		}).then(() => {
-			message.success(`Invite email sent to ${email}!`, 5)
-			emailRef.current?.focus()
-		})
-	}
-
 	if (error) {
 		return <div>{JSON.stringify(error)}</div>
 	}
@@ -124,113 +78,13 @@ const WorkspaceTeam = () => {
 							<AutoJoinForm label="Auto join" labelFirst />
 						</Authorization>
 					</div>
-					<Modal
-						destroyOnClose
-						centered
-						title="Invite Member"
-						visible={showModal}
-						width={600}
-						onCancel={toggleShowModal}
-					>
-						<form onSubmit={onSubmit}>
-							<p className={styles.boxSubTitle}>
-								Invite a team member to '
-								{`${data?.workspace?.name}`}' by entering an
-								email below.
-							</p>
-							<div className={styles.buttonRow}>
-								<Input
-									className={styles.emailInput}
-									placeholder="Email"
-									type="email"
-									name="invitedEmail"
-									autoFocus
-									value={email}
-									onChange={(e) => {
-										setEmail(e.target.value)
-									}}
-									addonAfter={
-										<Select
-											bordered={false}
-											value={newAdminRole}
-											options={(
-												Object.keys(
-													AdminRole,
-												) as (keyof typeof AdminRole)[]
-											).map((key) => {
-												const role = AdminRole[key]
-
-												return {
-													displayValue:
-														titleCaseString(role),
-													id: role,
-													value: role,
-												}
-											})}
-											onChange={setNewAdminRole}
-										/>
-									}
-								/>
-								<Button
-									trackingId="WorkspaceInviteMember"
-									type="primary"
-									className={clsx(
-										commonStyles.submitButton,
-										styles.inviteButton,
-									)}
-									htmlType="submit"
-								>
-									{sendLoading ? (
-										<CircularSpinner
-											style={{
-												fontSize: 18,
-												color: 'var(--text-primary-inverted)',
-											}}
-										/>
-									) : (
-										'Invite'
-									)}
-								</Button>
-							</div>
-						</form>
-						{sendInviteEmailData?.sendAdminWorkspaceInvite && (
-							<Alert
-								shouldAlwaysShow
-								trackingId="InviteAdminToWorkspaceConfirmation"
-								message="An invite email has been sent!"
-								type="success"
-								description={
-									<>
-										You can also share with them this link:{' '}
-										<span>
-											<CopyText
-												text={
-													sendInviteEmailData.sendAdminWorkspaceInvite
-												}
-												onCopyTooltipText="Copied invite link to clipboard!"
-												inline
-											/>
-										</span>
-									</>
-								}
-							/>
-						)}
-						<hr className={styles.hr} />
-						<p className={styles.boxSubTitle}>
-							Or share this link with them (this link expires{' '}
-							{moment(
-								data?.workspace_invite_links.expiration_date,
-							).fromNow()}
-							).
-						</p>
-						<CopyText
-							text={getWorkspaceInvitationLink(
-								data?.workspace_invite_links.secret || '',
-								workspace_id!,
-							)}
-							onCopyTooltipText="Copied invite link to clipboard!"
-						/>
-					</Modal>
+					<InviteMemberModal
+						workspaceId={workspace_id}
+						workspaceName={data?.workspace?.name}
+						workspaceInviteLinks={data?.workspace_invite_links}
+						showModal={showModal}
+						toggleShowModal={toggleShowModal}
+					/>
 					<Button
 						trackingId="WorkspaceTeamInviteMember"
 						type="primary"
