+++ frontend/src/pages/WorkspaceTeam/components/InviteMemberModal.module.scss.d.ts
@@ -1,0 +1,5 @@
+export const boxSubTitle: string
+export const buttonRow: string
+export const emailInput: string
+export const hr: string
+export const inviteButton: string
